refactor: Remove commented out code for post author in AdminDashboard
